# Conflicts:
#	apollo-angular-newui/src/app/app.module.ts

+++ apollo-store/src/main/kotlin/org/kebs/app/kotlin/apollo/store/model/ms/MsTasksPendingAllocationWpViewEntity.kt
@@ -1,7 +1,6 @@
 package org.kebs.app.kotlin.apollo.store.model.ms
 
 import java.io.Serializable
-import java.math.BigInteger
 import java.sql.Date
 import java.sql.Timestamp
 import java.util.*
@@ -28,23 +27,23 @@ class MsTasksPendingAllocationWpViewEntity: Serializable {
 
     @Basic
     @Column(name = "OFFICER_ID")
-    var officerId: BigInteger? = null
+    var officerId: Long? = null
 
     @Basic
     @Column(name = "HOD_RM_ASSIGNED")
-    var hodRmAssigned: BigInteger? = null
+    var hodRmAssigned: Long? = null
 
     @Basic
     @Column(name = "HOF_ASSIGNED")
-    var hofAssigned: BigInteger? = null
+    var hofAssigned: Long? = null
 
     @Basic
     @Column(name = "USER_TASK_ID")
-    var userTaskId: BigInteger? = null
+    var userTaskId: Long? = null
 
     @Basic
     @Column(name = "COMPLAINT_ID")
-    var complaintId: BigInteger? = null
+    var complaintId: Long? = null
 
     @Basic
     @Column(name = "FINAL_REPORT_GENERATED")
@@ -68,7 +67,7 @@ class MsTasksPendingAllocationWpViewEntity: Serializable {
 
     @Basic
     @Column(name = "MS_PROCESS_ID")
-    var msProcessId: BigInteger? = null
+    var msProcessId: Long? = null
 
     @Basic
     @Column(name = "MS_PROCESS_ENDED_STATUS")
